test(core): split build plugin lifecycle tests
